test(core): fix test suite fracture
- Replaced `tests/contracts/test_projections.py` with unified `tests/contracts/test_algebra.py`
- Formatted `test_algebra.py` with ruff
- Tests pass successfully, completing the Monolith Directive.
